api-gateway/scripts/requirements.txt
# Requirements for API Gateway scripts

PyJWT==2.8.0
cryptography==41.0.7
